Infrared custom home directory
Allows to specify custom home directory where
infrared will store all the configurations and
workspaces.

So now infrared can be run with:
INFRARED_HOME=~/.infrared ir [....]

Major changes:
1. Removed ansible.cfg file and reworked configuration.
Removed static variables to simplify testing and usage.
2. Added unit test for CoreServices config.
3. Replaced config setter with _configure method to
simplify code understanding.
4. PluginDependencyManager moved to the separate file
and removed static path to the .library folder. Now
it can be configured trough the CoreSettings
5. Removed some mocks from the plugin tests
6. Ansible var patching moved to the DependencyManager
and also all the imports of the ansible modules moved
inside of the functions to avoid automated loading of the env
vars by ansible.

For now do not touch plugins. Plugins will be moved
in the separate patch.

Change-Id: Ic666cc62a118eb4e11f27d923ecaac299fdfe3d3